crypt32: Don't use off_t for memory offsets.

+++ dlls/crypt32/encode.c
@@ -1215,7 +1215,7 @@ struct DERSetDescriptor
     DWORD                   cItems;
     const void             *items;
     size_t                  itemSize;
-    off_t                   itemOffset;
+    size_t                  itemOffset;
     CryptEncodeObjectExFunc encode;
 };
